* commit '7fd7d4bee42126234c4fcb98dd54bc9af750f808':
  fixing of several cppcheck warnings at several places

+++ tracking/trackFindingVXD/segmentNetwork/include/StaticSector.h
@@ -246,9 +246,9 @@ namespace Belle2 {
         char typeID = accessor[index].first;
         void* valuePtr = accessor[index].second;
         if (typeID == TBranchLeafType(double())) x = *((double*)valuePtr);
-        else if (typeID == TBranchLeafType(int())) x = *((int*)valuePtr);
-        else if (typeID == TBranchLeafType(float())) x = *((float*)valuePtr);
-        else if (typeID == TBranchLeafType(bool())) x = *((bool*)valuePtr);
+        else if (typeID == TBranchLeafType(int())) x = (double)(*((int*)valuePtr));
+        else if (typeID == TBranchLeafType(float())) x = (double)(*((float*)valuePtr));
+        else if (typeID == TBranchLeafType(bool())) x = (double)(*((bool*)valuePtr));
         else {
           B2FATAL("Unrecognized type : " << typeID);
         } // end else if
